Fix #3: Break up image sections from inside DashImage outward.
My primary motivation for this change was to address a bug that came to be
after issue #3 came up. I had linted the code with mypy and that revealed
to me that individual sections (e.g. text, gauge) were too tightly coupled
with the DashImage class. For example, gauges are required to record prior
states (last_gauge_value), and with them being tied to DashImage, I had to
delete the attribute at each new gauge's creation.

Ultimately, I felt that moving those methods to their respective section-based
classes could clean up the code and reduce overhead if/when a new image section
is added, especially now that there is a Section class that is to be inherited
by all subsequent image sections.